move unused functions
